Reorganized Files
Moved things to separate frontend, backend, and the old backend code
