--- Controller/api/UserController.php
@@ -1,79 +1,0 @@
-<?php
-class UserController extends BaseController
-{
-    /** 
-* "/user/list" Endpoint - Get list of users 
-*/
-    public function listAction()
-    {
-        $strErrorDesc = '';
-        $requestMethod = $_SERVER["REQUEST_METHOD"];
-        $arrQueryStringParams = $this->getQueryStringParams();
-        if (strtoupper($requestMethod) == 'GET') {
-            try {
-                $userModel = new UserModel();
-                $intLimit = 10;
-                if (isset($arrQueryStringParams['limit']) && $arrQueryStringParams['limit']) {
-                    $intLimit = $arrQueryStringParams['limit'];
-                }
-                $arrUsers = $userModel->getUsers($intLimit);
-                $responseData = json_encode($arrUsers);
-            } catch (Error $e) {
-                $strErrorDesc = $e->getMessage().'Something went wrong! Please contact support.';
-                $strErrorHeader = 'HTTP/1.1 500 Internal Server Error';
-            }
-        } else {
-            $strErrorDesc = 'Method not supported';
-            $strErrorHeader = 'HTTP/1.1 422 Unprocessable Entity';
-        }
-        // send output 
-        if (!$strErrorDesc) {
-            $this->sendOutput(
-                $responseData,
-                array('Content-Type: application/json', 'HTTP/1.1 200 OK')
-            );
-        } else {
-            $this->sendOutput(json_encode(array('error' => $strErrorDesc)), 
-                array('Content-Type: application/json', $strErrorHeader)
-            );
-        }
-    }
-    public function createAction()
-    {
-        $strErrorDesc = '';
-        $requestMethod = $_SERVER["REQUEST_METHOD"];
-    
-        if (strtoupper($requestMethod) == 'POST') {
-            $postData = json_decode(file_get_contents('php://input'), true);
-            $userModel = new UserModel();
-    
-            // Attempt to create the user
-            try {
-                $userModel->createUser($postData);
-    
-                // User was successfully created
-                $responseData = json_encode(array('message' => 'User created successfully'));
-            } catch (Exception $e) {
-                // Handle the case where user creation failed
-                $strErrorDesc = $e->getMessage();
-                $strErrorHeader = 'HTTP/1.1 500 Internal Server Error';
-            }
-        } else {
-            $strErrorDesc = 'Method not supported';
-            $strErrorHeader = 'HTTP/1.1 422 Unprocessable Entity';
-        }
-    
-        if (!$strErrorDesc) {
-            $this->sendOutput(
-                $responseData,
-                array('Content-Type: application/json', 'HTTP/1.1 201 Created')
-            );
-        } else {
-            $this->sendOutput(json_encode(array('error' => $strErrorDesc)),
-                array('Content-Type: application/json', $strErrorHeader)
-            );
-        }
-    }
-
-    // Add more functions as necessary
-}
