--- Model/Database.php
@@ -1,51 +1,0 @@
-<?php
-
-class Database
-{
-    protected $connection = null;
-
-    public function __construct()
-    {
-        try {
-            // Replace the placeholders with your actual database credentials
-            $this->connection = new mysqli('localhost', 'root', '', 'music_db');
-
-            if (mysqli_connect_errno()) {
-                throw new Exception("Could not connect to the database.");
-            }
-        } catch (Exception $e) {
-            throw new Exception($e->getMessage());
-        }
-    }
-
-    public function select($query = "", $params = [])
-    {
-        try {
-            $stmt = $this->executeStatement($query, $params);
-            $result = $stmt->get_result()->fetch_all(MYSQLI_ASSOC);
-            $stmt->close();
-            return $result;
-        } catch (Exception $e) {
-            throw new Exception($e->getMessage());
-        }
-        return false;
-    }
-
-    private function executeStatement($query = "", $params = [])
-    {
-        try {
-            $stmt = $this->connection->prepare($query);
-            if ($stmt === false) {
-                throw new Exception("Unable to create a prepared statement: " . $query);
-            }
-            if ($params) {
-                $stmt->bind_param($params[0], $params[1]);
-            }
-            $stmt->execute();
-            return $stmt;
-        } catch (Exception $e) {
-            throw new Exception($e->getMessage());
-        }
-    }
-}
-?>
