+++ backend/inc/config.php
@@ -1,0 +1,5 @@
+<?php
+define("DB_HOST", "localhost");
+define("DB_USERNAME", "root");
+define("DB_PASSWORD", "");
+define("DB_DATABASE_NAME", "music_db");
